fix(TagInput): moving files
