+++ packages/react-components/src/components/TagInput/index.ts
@@ -1,4 +1,4 @@
 export { TagInput } from './TagInput';
-export { EmailTagInput } from './EmailTagInput';
-export type { EmailTagInputProps } from './EmailTagInput';
+export { EmailTagInput } from './components/EmailTagInput';
+export type { EmailTagInputProps } from './components/EmailTagInput';
 export type { TagInputProps } from './types';
